Added a github deploy guide and resources folder for such things.
Figured this was a decent enough place for it and it is a PDF of a slideshow, because I like the way that breaks the steps into individual pages so people don't get overloaded reading it. Note: this is using the Windows GUI - some of the things should be transferable over to github for mac, but not all.