Converts some more slides over to reveal.js
I am saving a lot of images for later and screenshots so there are a lot of
comments that note what images need to be added throughout. Right now the
example html is done using &lt; &gt; - but there should be an easier way to
display the html. Maybe the markdown slides are the way to go for this?